requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cx-Freeze==6.1
decorator==4.4.2
discord.py==1.3.2
docopt==0.6.2
ffmpeg==1.4
ffprobe==0.5
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.7.12
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
jupyter-client==6.0.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
multidict==4.7.5
mysql-connector==2.2.9
parso==0.6.2
pickleshare==0.7.5
pipreqs==0.4.10
pluggy==0.13.1
prompt-toolkit==3.0.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.23.1
Pygments==2.6.1
pylint==2.4.4
PyNaCl==1.3.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytube3==9.6.4
pytz==2019.3
pywin32==227
pyzmq==19.0.0
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
websockets==8.1
winshell==0.6
wrapt==1.11.2
yarg==0.1.9
yarl==1.4.2
youtube-dl==2020.3.24
